+++ handlers/admin.py
@@ -11,9 +11,10 @@ from aiogram.utils.keyboard import InlineKeyboardBuilder
 from utils.create_bot import job_stores, bot, _
 from database.db_manage import get_user, update_user_sql, get_user_chats, get_chat_record, update_chat_sql
 # from handlers.client_handlers import ADMINS
-from keyboards.admin_kb import main_kb, reject_to_admin_btn, back_to_admin_btn, back_to_group_manage_btn, \
-    unblock_user_btn, block_user_btn, admin_menu_kb, back_my_channels_groups, back_my_channels_groups_kb, \
-    activate_ad_auction_kb
+from keyboards.admin_kb import reject_to_admin_btn, back_to_admin_btn, back_to_group_manage_btn, \
+    unblock_user_btn, block_user_btn, back_my_channels_groups, back_my_channels_groups_kb, \
+    activate_ad_auction_kb, admin_menu_kb
+from keyboards.client_kb import main_kb
 from utils.paypal import get_payment_status, create_order
 from utils.utils import  get_token_approval, payment_completed, \
     get_token_or_create_new
