+++ handlers/middleware.py
@@ -15,7 +15,7 @@ from utils.utils import translate_kb
 
 class HiddenUser(BaseMiddleware):
     async def on_process_message(self, message: types.Message, data):
-        from keyboards.admin_kb import main_kb
+        from keyboards.client_kb import main_kb
 
         if not message.from_user.username:
             await message.answer(text=_(
@@ -24,7 +24,7 @@ class HiddenUser(BaseMiddleware):
             raise CancelHandler()
 
     async def on_process_callback_query(self, query: types.CallbackQuery, data):
-        from keyboards.admin_kb import main_kb
+        from keyboards.client_kb import main_kb
         if not query.from_user.username:
             await query.message.answer(
                 text=_(
